Lots of cleanup, wstring porting, and now it lets you add a file.

+++ VFI18/ListView.h
@@ -1,5 +1,6 @@
 #pragma once
 
+//#include <stdint.h>
 #include "window.h"
 
 typedef std::shared_ptr<CWiseFile> spWiseFile;
@@ -13,14 +14,14 @@ public:
 
 	bool RegisterCreate(HINSTANCE hInstance, HWND hWnd);
 	unsigned int GetItemCount();
-	unsigned int GetTotalSize();
+	uint64_t GetTotalSize();
 	bool AddFile(std::wstring& strFile);
 
 private:
 	static LRESULT CALLBACK StaticSubClass(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam, UINT_PTR uIdSubclass, DWORD_PTR dwRefData);
 	bool InitColumns();
 	bool OnGetDispInfo(NMLVDISPINFO* plvdi);
-	DWORD m_qwSize;
+	uint64_t m_qwSize;
 
 public:
 	bool Resize(size_t sbHeight);
